MxShop/requirements.txt
djangorestframework==3.9.2
django==2.0.2
markdown==3.1
django-filter==2.1.0
pillow==6.0.0
pymysql==0.9.3
xadmin2==2.0.1
ipdb==0.12
django-guardian==1.5.0
coreapi==2.3.3
djangorestframework-jwt==1.11.0
pycryptodome==3.8.1